src/main/java/pl/pacinho/match/utils/ClassicFileReader.java
package pl.pacinho.match.utils;

import java.io.File;
import java.io.IOException;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.List;
import java.util.stream.Collectors;
import java.util.stream.Stream;

public class ClassicFileReader implements FileReader {

    @Override
    public List<String> readFile(File file) throws IOException {
        try (Stream<String> stream = Files.lines(Paths.get(file.getAbsolutePath()))) {
            return stream.collect(Collectors.toList());
        }
    }
}